fix(skills): address Greptile review — timeout passthrough, default_version mapping, E2E test location
- Pass timeout param to all GET-based skills HTTP handlers (content downloads)
- Add default_version to canonical Skill model so update_skill response preserves it
- Move OpenAI E2E tests from test_skills_api.py to test_skills_e2e.py (mock-only policy)
- Add 2 new unit tests for default_version canonical mapping

Refs: #22435
