+++ litellm/llms/custom_httpx/llm_http_handler.py
@@ -9524,7 +9524,7 @@ class BaseLLMHTTPHandler:
         )
 
         try:
-            response = sync_httpx_client.get(url=url, headers=headers)
+            response = sync_httpx_client.get(url=url, headers=headers, timeout=timeout)
         except Exception as e:
             raise self._handle_error(
                 e=e,
@@ -9570,7 +9570,7 @@ class BaseLLMHTTPHandler:
 
         try:
             response = await async_httpx_client.get(
-                url=url, headers=headers
+                url=url, headers=headers, timeout=timeout
             )
         except Exception as e:
             raise self._handle_error(
@@ -9761,7 +9761,7 @@ class BaseLLMHTTPHandler:
 
         try:
             response = sync_httpx_client.get(
-                url=url, headers=headers, params=query_params
+                url=url, headers=headers, params=query_params, timeout=timeout
             )
         except Exception as e:
             raise self._handle_error(
@@ -9810,7 +9810,7 @@ class BaseLLMHTTPHandler:
 
         try:
             response = await async_httpx_client.get(
-                url=url, headers=headers, params=query_params
+                url=url, headers=headers, params=query_params, timeout=timeout
             )
         except Exception as e:
             raise self._handle_error(
@@ -9869,7 +9869,7 @@ class BaseLLMHTTPHandler:
         )
 
         try:
-            response = sync_httpx_client.get(url=url, headers=headers)
+            response = sync_httpx_client.get(url=url, headers=headers, timeout=timeout)
         except Exception as e:
             raise self._handle_error(
                 e=e,
@@ -9915,7 +9915,7 @@ class BaseLLMHTTPHandler:
 
         try:
             response = await async_httpx_client.get(
-                url=url, headers=headers
+                url=url, headers=headers, timeout=timeout
             )
         except Exception as e:
             raise self._handle_error(
@@ -10081,7 +10081,7 @@ class BaseLLMHTTPHandler:
         )
 
         try:
-            response = sync_httpx_client.get(url=url, headers=headers)
+            response = sync_httpx_client.get(url=url, headers=headers, timeout=timeout)
         except Exception as e:
             raise self._handle_error(
                 e=e,
@@ -10127,7 +10127,7 @@ class BaseLLMHTTPHandler:
 
         try:
             response = await async_httpx_client.get(
-                url=url, headers=headers
+                url=url, headers=headers, timeout=timeout
             )
         except Exception as e:
             raise self._handle_error(
